// web/src/lib/api/generated/models/poskasir-internal-dto-update-order-item-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * POS Kasir API
 * POS Kasir API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { POSKasirInternalDtoCreateOrderItemOptionRequest } from './poskasir-internal-dto-create-order-item-option-request';

export interface POSKasirInternalDtoUpdateOrderItemRequest {
    'options'?: Array<POSKasirInternalDtoCreateOrderItemOptionRequest>;
    'product_id': string;
    'quantity': number;
}